moved community error and success code to consts pack
