+++ src/main/java/org/example/staystylish/domain/community/controller/LikeController.java
@@ -3,8 +3,8 @@ package org.example.staystylish.domain.community.controller;
 import lombok.RequiredArgsConstructor;
 import org.example.staystylish.common.dto.response.ApiResponse;
 import org.example.staystylish.common.security.UserPrincipal;
+import org.example.staystylish.domain.community.consts.CommunitySuccessCode;
 import org.example.staystylish.domain.community.dto.response.LikeResponse;
-import org.example.staystylish.domain.community.exception.CommunitySuccessCode;
 import org.example.staystylish.domain.community.service.LikeService;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.PathVariable;
